requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
pydantic>=2.0.0
